refact(scripts): refactored some script to reflect new main rules
